requirements.txt
alembic>=1.16.5
dependency_injector>=4.48.2
dotenv>=0.9.9
fastapi>=0.116.1
psycopg2-binary>=2.9.10
pydantic>=2.11.7
sqlalchemy>=2.0.43
uvicorn>=0.35.0